requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
Pillow==10.0.0
pix2tex==0.1.2
python-dotenv==1.0.0
matplotlib==3.7.2
